refactor: PresetSelectorコンポーネントの分離とロジックの整理
PresetSelectorをPresetInputとPresetDropdownに分離し、それぞれのコンポーネントに関連するロジックを移動しました。
また、リファクタリング後の動作確認のためのユニットテストを追加し、既存のE2Eテストが通過することを確認しました。
